tests/wpt/web-platform-tests/html/canvas/element/path-objects/2d.path.roundrect.radius.intersecting.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.roundrect.radius.intersecting.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.radius.intersecting.1</h1>
<p class="desc">Check that roundRects with intersecting corner arcs are rendered correctly.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Check that roundRects with intersecting corner arcs are rendered correctly.");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.roundRect(0, 0, 100, 50, [40, 40, 40, 40]);
ctx.fillStyle = '#0f0';
ctx.fill();
_assertPixel(canvas, 2,25, 0,255,0,255, "2,25", "0,255,0,255");
_assertPixel(canvas, 50,1, 0,255,0,255, "50,1", "0,255,0,255");
_assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");
_assertPixel(canvas, 50,48, 0,255,0,255, "50,48", "0,255,0,255");
_assertPixel(canvas, 97,25, 0,255,0,255, "97,25", "0,255,0,255");
_assertPixel(canvas, 1,1, 255,0,0,255, "1,1", "255,0,0,255");
_assertPixel(canvas, 98,1, 255,0,0,255, "98,1", "255,0,0,255");
_assertPixel(canvas, 1,48, 255,0,0,255, "1,48", "255,0,0,255");
_assertPixel(canvas, 98,48, 255,0,0,255, "98,48", "255,0,0,255");


});
</script>
